+++ src/app/notifications/page.jsx
@@ -1,16 +1,16 @@
 'use client';
 
-import React from 'react';
-import NotificationList from '../../components/notifications/NotificationList';
+import React, { useEffect, useState } from 'react';
+import NotificationItem from '../../components/NotificationItem';
 import Header from '../../components/layout/Header';
 import BottomNavigation from '../../components/layout/BottomNavigation';
 import styles from '../../styles/pages/notification.module.css';
 import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
 import axiosInstance from '../../api/axiosInstance';
 
 const NotificationsPage = () => {
   const router = useRouter();
+  const [notifications, setNotifications] = useState([]);
 
   useEffect(() => {
     const token = localStorage.getItem('accessToken');
@@ -27,7 +27,32 @@ const NotificationsPage = () => {
         localStorage.removeItem('accessToken');
         router.replace("/login");
       });
+
+    fetchNotifications();
   }, [router]);
+
+  const fetchNotifications = async () => {
+    try {
+      const response = await axiosInstance.get('/notifications');
+      const sorted = [...response.data].sort(
+        (a, b) => new Date(b.createdAt) - new Date(a.createdAt)
+      );
+      setNotifications(sorted);
+    } catch (error) {
+      console.error('알림을 불러오는데 실패했습니다:', error);
+    }
+  };
+
+  const handleNotificationClick = async (notification) => {
+    if (!notification.isRead) {
+      try {
+        await axiosInstance.post(`/notifications/${notification.id}/read`);
+        fetchNotifications(); // 읽음 처리 후 목록 갱신
+      } catch (error) {
+        console.error('알림 읽음 처리 실패:', error);
+      }
+    }
+  };
 
   return (
     <div className="mainContainer">
@@ -35,7 +60,20 @@ const NotificationsPage = () => {
       <div className="appContainer">
         <div className={styles.pageContainer}>
           <div className={styles.contentWrapper}>
-            <NotificationList />
+            <div style={{ display: 'flex', flexDirection: 'column', gap: '12px' }}>
+              {notifications.map((notification) => (
+                <NotificationItem
+                  key={notification.id}
+                  notification={notification}
+                  onClick={handleNotificationClick}
+                />
+              ))}
+              {notifications.length === 0 && (
+                <div style={{ textAlign: 'center', padding: '40px', color: '#6b7280' }}>
+                  알림이 없습니다
+                </div>
+              )}
+            </div>
           </div>
         </div>
       </div>
